added attack function and /combat routing in main.rs
#LABTROYALC-107

@author: Anton Schmalfuß

+++ clicker_backend/src/game_state.rs
@@ -101,7 +101,7 @@ impl GameState {
             }
             ClientMessages::UpgradeAutomationAmount => {
                 if self.ore as u64 >= upgrade_auto_amount_cost as u64 {
-                    self.ore -= upgrade_auto_amount_cost;
+                    self.ore -= upgrade_auto_amount_cost as f64;
                     self.auto_amount_level += 1;
                     upgrade_auto_amount_cost = self.auto_amount_level * 50;
                     ServerMessages::AutomationAmountUpgraded {
